requirements.txt
imageio
regex
requests
beautifulsoup4
Pillow
